修复实体报错

+++ core/src/main/java/org/geysermc/geyser/entity/type/Entity.java
@@ -141,8 +141,8 @@ public class Entity {
      */
     protected void initializeMetadata() {
         dirtyMetadata.put(EntityData.SCALE, 1f);
-        dirtyMetadata.put(EntityData.COLOR, 0);
-        dirtyMetadata.put(EntityData.MAX_AIR_SUPPLY, getMaxAir());
+        dirtyMetadata.put(EntityData.COLOR, (byte) 0);
+        dirtyMetadata.put(EntityData.MAX_AIR_SUPPLY, (short) getMaxAir());
         setDimensions(Pose.STANDING);
         setFlag(EntityFlag.HAS_GRAVITY, true);
         setFlag(EntityFlag.HAS_COLLISION, true);
